refact: extract prepareSafeTransaction function
